requirements.txt
streamlit==1.44.1
weaviate-client==3.26.7
python-dotenv==1.1.0
pypdf==3.8.1  # replace PyPDF2 with pypdf
python-docx==1.1.2
langchain-community==0.3.24
groq==0.23.0
